+++ enterprise/api_client/discovery.py
@@ -5,6 +5,7 @@ Utilities to get details from the course catalog API.
 from __future__ import absolute_import, unicode_literals
 
 from logging import getLogger
+from sys import getsizeof
 
 from edx_rest_api_client.client import EdxRestApiClient
 from edx_rest_api_client.exceptions import SlumberBaseException
@@ -164,6 +165,11 @@ class CourseCatalogApiClient(object):
                     content_filter_query,
                     query_params,
                     traverse_pagination
+                )
+                LOGGER.info(
+                    'ENT-2489 | Response from content_filter_query %s is %d bytes long.',
+                    content_filter_query,
+                    getsizeof(response)
                 )
                 cache.set(cache_key, response, settings.ENTERPRISE_API_CACHE_TIMEOUT)
             else: