+++ integrated_channels/xapi/utils.py
@@ -10,8 +10,8 @@ import logging
 
 import six
 
+from enterprise.tpa_pipeline import get_user_social_auth
 from integrated_channels.xapi.client import EnterpriseXAPIClient
-from integrated_channels.xapi.serializers import CourseInfoSerializer, LearnerInfoSerializer
 from integrated_channels.xapi.statements.learner_course_completion import LearnerCourseCompletionStatement
 from integrated_channels.xapi.statements.learner_course_enrollment import LearnerCourseEnrollmentStatement
 
@@ -26,25 +26,19 @@ def send_course_enrollment_statement(lrs_configuration, course_enrollment):
          lrs_configuration (XAPILRSConfiguration): XAPILRSConfiguration instance where to send statements.
          course_enrollment (CourseEnrollment): Course enrollment object.
     """
-    user_details = LearnerInfoSerializer(
-        course_enrollment.user,
-        context={'enterprise_customer': lrs_configuration.enterprise_customer}
-    )
-    course_details = CourseInfoSerializer(
-        course_enrollment.course,
-        context={'site': lrs_configuration.enterprise_customer.site}
-    )
+    user = course_enrollment.user
     LOGGER.info(
         'Sending course enrollment to xAPI for user: {username} for course: {course_key}'.format(
-            username=course_enrollment.user.username,
+            username=user.username,
             course_key=six.text_type(course_enrollment.course.id)
         )
     )
+
+    user_social_auth = get_user_social_auth(user, lrs_configuration.enterprise_customer)
     statement = LearnerCourseEnrollmentStatement(
-        course_enrollment.user,
+        user,
+        user_social_auth,
         course_enrollment.course,
-        user_details.data,
-        course_details.data,
     )
     EnterpriseXAPIClient(lrs_configuration).save_statement(statement)
 
@@ -59,14 +53,6 @@ def send_course_completion_statement(lrs_configuration, user, course_overview, c
          course_overview (CourseOverview): Course over view object containing course details.
          course_grade (CourseGrade): course grade object.
     """
-    user_details = LearnerInfoSerializer(
-        user,
-        context={'enterprise_customer': lrs_configuration.enterprise_customer}
-    )
-    course_details = CourseInfoSerializer(
-        course_overview,
-        context={'site': lrs_configuration.enterprise_customer.site}
-    )
     LOGGER.info(
         'Sending course completion to xAPI for user: {username}, course: {course_key} with {percentage}%'.format(
             username=user.username if user else '',
@@ -74,11 +60,11 @@ def send_course_completion_statement(lrs_configuration, user, course_overview, c
             percentage=course_grade.percent * 100
         )
     )
+    user_social_auth = get_user_social_auth(user, lrs_configuration.enterprise_customer)
     statement = LearnerCourseCompletionStatement(
         user,
+        user_social_auth,
         course_overview,
-        user_details.data,
-        course_details.data,
         course_grade,
     )
     EnterpriseXAPIClient(lrs_configuration).save_statement(statement)