+++ tests/test_integrated_channels/test_xapi/test_statements/test_course_completion.py
@@ -14,7 +14,7 @@ from pytest import mark
 
 from integrated_channels.xapi.constants import X_API_ACTIVITY_COURSE, X_API_VERB_COMPLETED
 from integrated_channels.xapi.statements.learner_course_completion import LearnerCourseCompletionStatement
-from test_utils import TEST_COURSE, TEST_UUID, factories
+from test_utils import factories
 
 
 @mark.django_db
@@ -28,27 +28,14 @@ class TestLearnerCourseCompletionStatement(unittest.TestCase):
         faker = FakerFactory.create()
 
         self.user = factories.UserFactory()
+        self.mock_social_auth = Mock(provider='tpa-saml', uid='default:edxsso')
         # pylint: disable=no-member
-        self.course_overview = Mock(display_name=faker.text(max_nb_chars=25), short_description=faker.text())
+        self.course_overview = Mock(
+            id='course-v1:edX+DemoX+Demo_Course',
+            display_name=faker.text(max_nb_chars=25),
+            short_description=faker.text()
+        )
         self.course_grade = Mock(percent=0.80, passed=True)
-
-        self.user_details = {
-            'enrollment_created_timestamp': '2018-06-25T04:45:32.642094',
-            'enterprise_user_id': 1,
-            'lms_user_id': 2,
-            'enterprise_sso_uid': TEST_UUID,
-            'user_account_creation_date': '2017-09-27T02:51:32.612293',
-            'user_email': 'user@example.com',
-            'user_username': 'test_user',
-            'user_country_code': 'PK',
-            'user_current_enrollment_mode': 'verified',
-        }
-        self.course_details = {
-            'course_id': TEST_COURSE,
-            'course_title': self.course_overview.display_name,
-            'course_duration': '6 Months',
-            'course_min_effort': '3 days per week',
-        }
 
         self.expected = {
             'verb': {
@@ -58,11 +45,12 @@ class TestLearnerCourseCompletionStatement(unittest.TestCase):
             'version': '1.0.1',
             'actor': {
                 'mbox': 'mailto:{email}'.format(email=self.user.email),
-                'name': self.user.username,
+                'name': 'edxsso',
                 'objectType': 'Agent'
             },
             'object': {
                 'definition': {
+                    'type': X_API_ACTIVITY_COURSE,
                     'description': {
                         'en-US': self.course_overview.short_description
                     },
@@ -70,7 +58,7 @@ class TestLearnerCourseCompletionStatement(unittest.TestCase):
                         'en-US': self.course_overview.display_name
                     }
                 },
-                'id': X_API_ACTIVITY_COURSE,
+                'id': self.course_overview.id,
                 'objectType': 'Activity'
             },
             'result': {
@@ -83,12 +71,6 @@ class TestLearnerCourseCompletionStatement(unittest.TestCase):
                 'success': True,
                 'completion': True,
             },
-            'context': {
-                'extensions': {
-                    'http://id.tincanapi.com/extension/course-details': self.course_details,
-                    'http://id.tincanapi.com/extension/user-details': self.user_details
-                }
-            },
         }
 
     def test_statement(self):
@@ -97,9 +79,8 @@ class TestLearnerCourseCompletionStatement(unittest.TestCase):
         """
         statement = LearnerCourseCompletionStatement(
             self.user,
+            self.mock_social_auth,
             self.course_overview,
-            self.user_details,
-            self.course_details,
             self.course_grade,
         )
-        assert json.loads(statement.to_json()) == self.expected
+        self.assertDictEqual(json.loads(statement.to_json()), self.expected)